OCR App v2/requirements.txt
customtkinter>=5.2.2
tkinterdnd2>=0.3.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
